Change the file structure of the `server/controllers/` The `controllers/` now uses the index as a directory catalog to other files. Added comments to index.js and client-controller.js as sample JDocs for the project.  Changes to be committed: 	new file:   server/controllers/client-controller.js 	new file:   server/controllers/dashboard-controller.js 	modified:   server/controllers/index.js 	new file:   server/controllers/user-controller.js
closes #86
