+++ server/controllers/index.js
@@ -1,49 +1,11 @@
+'use strict';
 
-var db = require('../../db');
-
-
-
-
+/**
+ * A map of all controllers. Requiring `controllers` will connect you to all controllers in the directory.
+ * @type {Object}
+ */
 module.exports = {
-	dashboard: {
-		get: function(req,res){
-			var userid = req.params.user_id
-			db.getAllClients(userid, function(err,result){
-				if(err){
-					throw err;
-				} else {
-					res.json(result);
-				}
-
-			});
-			// This is where we call the DB function for getting dashboard
-
-		}
-	},
-	user: {
-		post: function(req,res){
-			// This is where we call the DB function for creating new user
-			// TO DO: Make sure you send a response
-		}	
-	},
-	client: {
-		get: function(req,res){
-			var userid = req.params.user_id;
-			var clientid = req.params.client_id;
-			// This is where we call the DB function for getting particualr client
-		},
-		post: function(req,res){
-			var userid = req.params.user_id;
-			// This is where we call the DB function for creating a new client
-		},
-		put: function(req,res){
-			var user_id = req.params.user_id;
-			// This is where we call the DB function for updating a client
-		}
-	}
-
+	dashboard: require('./dashboard-controller'),
+  user: require('./user-controller'),
+  client: require('/client-controller')
 }
-
-
-
-
