+++ server/controllers/user-controller.js
@@ -1,0 +1,9 @@
+'use strict';
+var db = require('../../db');
+
+module.exports = {
+  post: function(req,res){
+      // This is where we call the DB function for creating new user
+      // TO DO: Make sure you send a response
+    }
+}
